Rename moveDependencies to copyDependencies and specify version of `fs-extra` to install
