+++ lib/copy-dependencies.js
