+++ lib/demeteorizer.js
@@ -2,7 +2,7 @@ const Hoek = require('hoek')
 
 const Build = require('./build')
 const Install = require('./install')
-const MoveDependencies = require('./move-dependencies')
+const CopyDependencies = require('./copy-dependencies')
 const UpdatePackage = require('./update-package')
 
 module.exports = (options, done) => {
@@ -14,8 +14,8 @@ module.exports = (options, done) => {
     Build(options, (err) => {
       if (err) return done(err)
 
-      MoveDependencies(options, (err) => {
-        if (err) return done(err)
+      CopyDependencies(options, (err) => {
+        if (err) return done(err);
 
         UpdatePackage(options, done)
       })
