+++ package.json
@@ -25,7 +25,7 @@
   "dependencies": {
     "commander": "2.9.0",
     "cross-spawn": "3.0.1",
-    "fs-extra": "^0.30.0",
+    "fs-extra": "0.30.0",
     "hoek": "4.x.x",
     "semver": "5.1.0"
   },
